\name{QNB-package}
\alias{QNB-package}
\alias{QNB}
\docType{package}
\title{
Differential Methylation Analysis with Quadratic-Negative-Binomial Model
}
\description{
It  has been developed for differential methylation analysis. Estimate variance and mean dependence in count data from MeRIP-seq and test for differential methylation based on a model using quadratic-negative-binomial distribution. Please don't hesitate to <liulian19860905@163.com> if you have any problem. The inputs of the main function are four reads count matrix for IP samples of two conditions and Input samples of two conditions.
}
\details{
\tabular{ll}{
Package: \tab QNB\cr
Type: \tab Package\cr
Version: \tab 0.99.2\cr
Date: \tab 2016-10-13\cr
License: \tab GPL-2\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Lian Liu, PhD 
Maintainer: <liulian19860905@163.com>
}
\keyword{ differential methylation,quadratic-negative-binomial model }

\examples{
library(QNB)
?QNB
}