\name{QNB-package}
\alias{QNB-package}
\alias{QNB}
\docType{package}
\title{
Differential methylation analysis with quadratic-negative-binomial model 
}
\description{
The package is developed for differential methylation analysis. Estimate variance and mean dependence in count data from MeRIP-seq and test for differential
methylation based on a model using quadratic-negative-binomial distribution.
}
\details{
\tabular{ll}{
Package: \tab QNB\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-07-16\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Lian Liu, PhD 
Maintainer: <liulian19860905@163.com>
}
\keyword{ differential methylation,quadratic-negative-binomial model }

\examples{
library(QNB)
?QNB
}